EROPSPT-399: Reset final remove date when reissuing AED (#322)
* EROPSPT-399: Reset final remove date when reissuing AED

* EROPSPT-391: Retain photos used by retained AEDs

+++ src/main/kotlin/uk/gov/dluhc/printapi/service/AedDataRetentionService.kt
@@ -7,6 +7,7 @@ import uk.gov.dluhc.printapi.database.entity.SourceType
 import uk.gov.dluhc.printapi.database.repository.AnonymousElectorDocumentRepository
 import uk.gov.dluhc.printapi.database.repository.AnonymousElectorDocumentRepositoryExtensions.findPendingRemovalOfFinalRetentionData
 import uk.gov.dluhc.printapi.database.repository.AnonymousElectorDocumentRepositoryExtensions.findPendingRemovalOfInitialRetentionData
+import uk.gov.dluhc.printapi.database.repository.AnonymousElectorDocumentRepositoryExtensions.shouldRetainPhoto
 import uk.gov.dluhc.printapi.mapper.SourceTypeMapper
 import uk.gov.dluhc.printapi.messaging.models.ApplicationRemovedMessage
 
@@ -65,7 +66,14 @@ class AedDataRetentionService(
         with(anonymousElectorDocumentRepository.findPendingRemovalOfFinalRetentionData(sourceType = sourceType)) {
             logger.info { "Found $size Anonymous Elector Documents with sourceType $sourceType to remove" }
             forEach {
-                s3AccessService.removeDocument(it.photoLocationArn)
+                val shouldRetainPhoto = anonymousElectorDocumentRepository.shouldRetainPhoto(
+                    photoLocationArn = it.photoLocationArn
+                )
+
+                if (!shouldRetainPhoto) {
+                    s3AccessService.removeDocument(it.photoLocationArn)
+                }
+
                 anonymousElectorDocumentRepository.deleteById(it.id!!)
             }
         }
